Refactor work to eliminate depricated `Quota`
Direct access to NumCpu, RamMB, and DiskMb are
deprecated and set to be remove in Aurora 0.20.0
(See https://issues.apache.org/jira/browse/AURORA-1707)

There does appear to be an issue when `NumGPU` and
`NamedPort` are set that the thrift clients errors
and fails. I will continue to investigate this.